Fixed a fault with key & value tokens when quoted
